fix(work): close review lifecycle and lint gaps
